requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blobfile==3.1.0
boto3==1.40.61
botocore==1.40.61
certifi==2025.11.12
charset-normalizer==3.4.4
chess==1.11.2
chz==0.3.0
click==8.2.1
datasets==4.4.1
debugpy==1.8.17
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
filelock==3.20.0
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ijson==3.4.0.post0
inspect_ai==0.3.146
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2sympy2_extended==1.10.2
linkify-it-py==2.0.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
math-verify==0.8.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.5
openai==2.8.1
packaging==25.0
pandas==2.3.3
pathlib_abc==0.5.2
platformdirs==4.5.0
ply==3.11
propcache==0.4.1
psutil==7.1.3
pyarrow==22.0.0
pycryptodomex==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
s3fs==2025.9.0
s3transfer==0.14.0
safetensors==0.7.0
scipy==1.16.3
semver==3.0.4
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
TextArena==0.7.4
textual==6.6.0
tinker==0.5.0
-e git+https://github.com/thinking-machines-lab/tinker-cookbook.git@0adfc25b7e1ea99d3ec4aab311516fab8c5e5120#egg=tinker_cookbook
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.3.6
urllib3==2.5.0
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
